requirements.txt
requests
pandas
snowflake-connector-python
apache-airflow
python-dotenv